Moved get_entity function to utilities.py
